Elemental 7 gRPC update

+++ elemental/util.go
@@ -77,7 +77,7 @@ func (e *Elemental) addCombo(elem1 string, elem2 string, out string) error {
 	return nil
 }
 
-func max(val1 int, val2 int) int {
+func max(val1 int64, val2 int64) int64 {
 	if val1 > val2 {
 		return val1
 	}
